requirements.txt
albumentations               1.3.0
einops                       0.6.0
h5py                         3.9.0
imageio                      2.25.1
matplotlib                   3.7.0
numpy                        1.24.1
opencv-python                4.7.0.68
Pillow                       9.4.0
scikit-image                 0.19.3
scikit-learn                 1.2.1
scipy                        1.10.0
timm                         0.6.12
torch                        1.13.1
torchvision                  0.14.1
tqdm                         4.64.1